backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0